change the location of function "Renew_Container"
